# .development/specs/001-theme-park-tracker/contracts/api.yaml
openapi: 3.0.3
info:
  title: Theme Park Downtime Tracker API
  description: |
    REST API for the Theme Park Hall of Shame downtime tracking system.

    This API provides endpoints for retrieving park downtime rankings, ride performance statistics,
    and real-time wait times for North American theme parks. Data is collected from Queue-Times.com
    every 10 minutes and aggregated into daily, weekly, monthly, and yearly statistics.

    **Key Features**:
    - Park rankings by total downtime hours
    - Individual ride performance tracking
    - Real-time wait time monitoring
    - Flexible time period filtering (live, today, 7 days, 30 days)
    - Disney & Universal park filtering

    **Data Attribution**: All data powered by Queue-Times.com
  version: 1.0.0
  contact:
    name: Theme Park Hall of Shame
    url: https://themeparkwaits.com
  license:
    name: Proprietary
servers:
  - url: https://api.themeparkwaits.com/v1
    description: Production server
  - url: http://localhost:5000/v1
    description: Local development server
...
paths:
  /parks/downtime:
    get:
      tags:
        - parks
      summary: Get park downtime rankings
      description: |
        Returns parks ranked by total downtime hours for the specified time period.
        Supports filtering by park operator (Disney & Universal vs all parks).

        **Performance**: Query executes in <50ms for daily data, <100ms for weekly/monthly data.

        **Business Logic**:
        - Downtime calculated only during park operating hours (excludes overnight closures)
        - Trend percentage compares current period to previous equivalent period
        - Affected rides count shows currently closed rides from most recent snapshot
      operationId: getParkDowntimeRankings
      parameters:
        - name: period
          in: query
          description: |
            Time period for ranking calculation:
            - `live`: Instantaneous snapshot (rides down RIGHT NOW)
            - `today`: Cumulative from midnight Pacific to now
            - `7days`: Rolling 7-day aggregate
            - `30days`: Rolling 30-day aggregate
          required: false
          schema:
            type: string
            enum:
              - live
              - today
              - 7days
              - 30days
            default: live
        - name: filter
          in: query
          description: Filter parks by operator
          required: false
          schema:
            type: string
            enum:
              - disney-universal
              - all-parks
            default: all-parks
        - name: limit
          in: query
          description: Maximum number of parks to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Successful response with park rankings
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  period:
                    type: string
                    example: today
                  filter:
                    type: string
                    example: all-parks
                  aggregate_stats:
                    $ref: '#/components/schemas/AggregateStats'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ParkDowntimeRanking'
              examples:
                todayAllParks:
                  summary: Today's rankings for all parks
                  value:
                    success: true
                    period: today
                    filter: all-parks
                    aggregate_stats:
                      total_parks_tracked: 85
                      peak_downtime_hours: 12.5
                      currently_down_rides: 23
                    data:
                      - rank: 1
                        park_id: 16
                        park_name: Magic Kingdom
                        location: Orlando, FL
                        total_downtime_hours: 12.5
                        affected_rides_count: 5
                        uptime_percentage: 89.2
                        trend_percentage: 15.3
                        queue_times_url: https://queue-times.com/parks/16
                      - rank: 2
                        park_id: 17
                        park_name: Universal Studios Florida
                        location: Orlando, FL
                        total_downtime_hours: 8.7
                        affected_rides_count: 3
                        uptime_percentage: 92.5
                        trend_percentage: -5.2
                        queue_times_url: https://queue-times.com/parks/17
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /parks/{parkId}/details:
    get:
      tags:
        - parks
      summary: Get detailed park information
      description: |
        Returns detailed information about a specific park including location,
        operator, current operational status, and summary statistics.
      operationId: getParkDetails
      parameters:
        - name: parkId
          in: path
          description: Park identifier
          required: true
          schema:
            type: integer
            example: 16
        - name: period
          in: query
          description: |
            Time period for statistics:
            - `live`: Instantaneous snapshot
            - `today`: Cumulative from midnight Pacific to now
            - `7days`: Rolling 7-day aggregate
            - `30days`: Rolling 30-day aggregate
          required: false
          schema:
            type: string
            enum:
              - live
              - today
              - 7days
              - 30days
            default: live
      responses:
        '200':
          description: Successful response with park details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ParkDetails'
              examples:
                magicKingdom:
                  summary: Magic Kingdom details
                  value:
                    success: true
                    data:
                      park_id: 16
                      name: Magic Kingdom
                      location:
                        city: Orlando
                        state: FL
                        country: US
                      coordinates:
                        latitude: 28.417663
                        longitude: -81.581213
                      operator: Disney
                      timezone: America/New_York
                      total_rides_tracked: 45
                      current_status:
                        rides_open: 42
                        rides_closed: 3
                        avg_wait_time: 32.5
                        max_wait_time: 90
                      statistics:
                        period: today
                        total_downtime_hours: 12.5
                        avg_uptime_percentage: 89.2
                        rides_with_downtime: 5
                      queue_times_url: https://queue-times.com/parks/16
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    ParkDowntimeRanking:
      type: object
      description: Park ranked by weighted downtime score
      required:
        - rank
        - park_id
        - park_name
        - location
        - total_downtime_hours
        - weighted_downtime_score
        - score_percentage
        - tier_distribution
        - affected_rides_count
        - uptime_percentage
        - trend_percentage
        - queue_times_url
      properties:
        rank:
          type: integer
          description: Ranking position (1 = highest weighted downtime score)
          example: 1
        park_id:
          type: integer
          description: Park identifier
          example: 16
        park_name:
          type: string
          description: Park name
          example: Magic Kingdom
        location:
          type: string
          description: City and state/province formatted as "City, State"
          example: Orlando, FL
        total_downtime_hours:
          type: number
          format: float
          description: Raw total downtime hours for the period (unweighted)
          example: 12.5
        weighted_downtime_score:
          type: number
          format: float
          description: Weighted downtime score = total_weighted_downtime_hours / total_park_weight (0.0 to 1.0+)
          example: 0.7654
        score_percentage:
          type: number
          format: float
          description: Weighted score as percentage (0-100%) representing % of weighted capacity offline
          example: 76.54
        tier_distribution:
          type: object
          description: Count of rides by tier (Tier 1=major, Tier 2=standard, Tier 3=minor)
          required:
            - tier1_count
            - tier2_count
            - tier3_count
            - total_park_weight
          properties:
            tier1_count:
              type: integer
              description: Number of Tier 1 rides (3x weight)
              example: 11
            tier2_count:
              type: integer
              description: Number of Tier 2 rides (2x weight)
              example: 18
            tier3_count:
              type: integer
              description: Number of Tier 3 rides (1x weight)
              example: 12
            total_park_weight:
              type: integer
              description: Sum of all ride weights (Tier1*3 + Tier2*2 + Tier3*1)
              example: 81
        affected_rides_count:
          type: integer
          description: Number of rides currently closed (from most recent snapshot)
          example: 5
        uptime_percentage:
          type: number
          format: float
          description: Average uptime percentage during operating hours (unweighted)
          example: 89.2
        trend_percentage:
          type: number
          format: float
          description: Percentage change vs previous equivalent period (positive = worse, negative = better)
          example: 15.3
        queue_times_url:
          type: string
          format: uri
          description: Link to park details on Queue-Times.com
          example: https://queue-times.com/parks/16
    ParkDetails:
      type: object
      description: Detailed park information
      required:
        - park_id
        - name
        - location
        - operator
        - timezone
        - total_rides_tracked
        - current_status
        - statistics
        - queue_times_url
      properties:
        park_id:
          type: integer
          example: 16
        name:
          type: string
          example: Magic Kingdom
        location:
          type: object
          properties:
            city:
              type: string
              example: Orlando
            state:
              type: string
              example: FL
            country:
              type: string
              example: US
        coordinates:
          type: object
          nullable: true
          properties:
            latitude:
              type: number
              format: float
              example: 28.417663
            longitude:
              type: number
              format: float
              example: -81.581213
        operator:
          type: string
          example: Disney
        timezone:
          type: string
          description: IANA timezone identifier
          example: America/New_York
        total_rides_tracked:
          type: integer
          example: 45
        current_status:
          type: object
          properties:
            rides_open:
              type: integer
              example: 42
            rides_closed:
              type: integer
              example: 3
            avg_wait_time:
              type: number
              format: float
              nullable: true
              example: 32.5
            max_wait_time:
              type: integer
              nullable: true
              example: 90
        statistics:
          type: object
          properties:
            period:
              type: string
              example: today
            total_downtime_hours:
              type: number
              format: float
              example: 12.5
            avg_uptime_percentage:
              type: number
              format: float
              example: 89.2
            rides_with_downtime:
              type: integer
              example: 5
        queue_times_url:
          type: string
          format: uri
          example: https://queue-times.com/parks/16
...
    AggregateStats:
      type: object
      description: Top-level aggregate statistics
      required:
        - total_parks_tracked
        - peak_downtime_hours
        - currently_down_rides
      properties:
        total_parks_tracked:
          type: integer
          description: Total number of parks in the filtered dataset
          example: 85
        peak_downtime_hours:
          type: number
          format: float
          description: Maximum downtime hours among all parks
          example: 12.5
        currently_down_rides:
          type: integer
          description: Total number of rides currently closed across all parks
          example: 23
...
    Error:
      type: object
      description: Error response
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: INVALID_PARAMETER
            message:
              type: string
              description: Human-readable error message
              example: "Invalid period parameter. Must be one of: today, 7days, 30days"
            details:
              type: object
              nullable: true
              description: Additional error details
              additionalProperties: true
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalidPeriod:
              summary: Invalid period parameter
              value:
                success: false
                error:
                  code: INVALID_PARAMETER
                  message: "Invalid period parameter. Must be one of: today, 7days, 30days"
                  details:
                    parameter: period
                    provided_value: yesterday
            invalidFilter:
              summary: Invalid filter parameter
              value:
                success: false
                error:
                  code: INVALID_PARAMETER
                  message: "Invalid filter parameter. Must be one of: disney-universal, all-parks"
                  details:
                    parameter: filter
                    provided_value: six-flags
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            parkNotFound:
              summary: Park not found
              value:
                success: false
                error:
                  code: RESOURCE_NOT_FOUND
                  message: Park with ID 99999 not found
                  details:
                    resource_type: park
                    resource_id: 99999
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            databaseError:
              summary: Database connection error
              value:
                success: false
                error:
                  code: INTERNAL_ERROR
                  message: An unexpected error occurred. Please try again later.
                  details: null
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Required for all endpoints.

        **Rate Limits**:
        - 100 requests per hour per API key
        - 1000 requests per day per API key

        **Obtaining an API Key**:
        Contact the administrator or register at https://themeparkwaits.com/api-access

        **Usage**:
        Include the API key in the request header:
        ```
        X-API-Key: your_api_key_here
        ```
security:
  - ApiKeyAuth: []